Loki: Move log context to separate provider
